mission accomplished 🔥🚀🔥

+++ CodeChef/22_November_2023/Alter_Ego.cpp
@@ -14,18 +14,18 @@ typedef pair<int,int>pii;
 #define forl(ty,var,str,end) for(ty var=str; var<end; var++)
 # define FAST ios_base :: sync_with_stdio (false) ; cin.tie(0) ; cout.tie(0)
 void solve(){
-    ll n;cin>>n;
-    ll a[n],even=0,odd=0;
-    for(ll i=0;i<n;i++)cin>>a[i];
-    for(ll i=0;i<n;i++){
+    int n;cin>>n;
+    int a[n],even=0,odd=0;
+    for(int i=0;i<n;i++)cin>>a[i];
+    for(int i=0;i<n;i++){
         if(a[i]&1)odd++;
         else even++;
     }
     if((even&1)||(odd&1))cout<<-1<<nl;
     else{
         bool ok=0;
-        vector<ll>ans;
-        for(ll i=0;i<n;i+=2){
+        vector<int>ans;
+        for(int i=0;i<n;i+=2){
             if((a[i]%2==0&&a[i+1]%2==0)||(a[i]%2!=0&&a[i+1]%2!=0)){
                 ans.push_back((a[i]+a[i+1])/2);
                 ans.push_back(abs(a[i]-(a[i]+a[i+1])/2));
@@ -35,7 +35,7 @@ void solve(){
         else{
             sort(all(ans));
             reverse(all(ans));
-            for(ll val:ans)cout<<val<<" ";
+            for(int val:ans)cout<<val<<" ";
             cout<<nl;
         }
     }
